Bayesian logistic regression models with correlated measurement errors

This function implements the Bayesian logistic regression model with correlated measurement error of covariate(s). Denote \(Y_i\) be a binary response, \(X_i\) be a \(q\times 1\) covariate of \(i\)th observation that is subject to measurement error, and \(Z_i\) be a \(p\times 1\) covariate without measurement error. The likelihood model is Bayesian logistic regression, $$logit(Pr(Y_i = 1)) = \beta_0 + X_i^\top \beta_x + Z_i^\top \beta_z, \quad i=1,\dots,n, $$ independently, and correlated measurement error of \(X_i, i=1,\dots,n\) is incorporated into the model as a multivariate normal prior. For example when \(q=1\), we have \(n-\)dimensional multivariate normal prior $$(X_1,\dots,X_n)\sim N_n(\mu_X, Q_X^{-1}).$$ Also, we consider semiconjugate priors for regression coefficients; $$\beta_0 \sim N(0, V_\beta), \quad \beta_{x,j} \stackrel{iid}{\sim} N(0, V_\beta), \quad \beta_{z,k} \stackrel{iid}{\sim} N(0, V_\beta)$$ The function blogireg_me() implements the Gibbs sampler for posterior inference using Polya-Gamma augmentation. Most importantly, it allows sparse matrix input for \(Q_X\) for scalable computation.

Usage

blogireg_me(
  Y,
  X_mean,
  X_prec,
  Z,
  nburn = 2000,
  nsave = 2000,
  nthin = 5,
  prior = NULL,
  saveX = F
)

Arguments

Y

n by 1 matrix, binary responses

X_mean

n by 1 matrix of mean of X \(\mu_X\), or list of n by 1 matrices of length q.

X_prec

n by n matrix precision matrix of X \(Q_X\), or list of n by n matrices of length q. Support sparse matrix format from Matrix package.

Z

n by p matrix, covariates without measurement error

nburn

integer, burn-in iteration

nsave

integer, number of posterior samples

nthin

integer, thin-in rate

prior

list of prior parameters, default is var_beta = 100,a_Y = 0.01, b_Y = 0.01

saveX

logical, save X or not

Value

list of (1) posterior, the (nsave)x(q+p) matrix of posterior samples as a coda object, (2) cputime, cpu time taken in seconds, and optionally (3) X_save, posterior samples of X

Examples

if (FALSE) {
data(ozone)
data(health_sim)
library(bspme)
data(ozone)
data(health_sim)
library(fields)
# exposure mean and covariance at health subject locations with 06/18/1987 data (date id = 16)
# using Gaussian process with prior mean zero and exponential covariance kernel
# with fixed range 300 (in miles) and stdev 15 (in ppb)

ozone16 = ozone[ozone$date_id==16,]

Dxx = rdist.earth(cbind(ozone16$coords_lon, ozone16$coords_lat))
Dyy = rdist.earth(cbind(health_sim$coords_y_lon, health_sim$coords_y_lat))
Dxy = rdist.earth(cbind(ozone16$coords_lon, ozone16$coords_lat),
                  cbind(health_sim$coords_y_lon, health_sim$coords_y_lat))

Kxx = Exponential(Dxx, range = 300, phi=15^2)
Kyy = Exponential(Dyy, range = 300, phi=15^2)
Kxy = Exponential(Dxy, range = 300, phi=15^2)

X_mean = t(Kxy) %*% solve(Kxx, ozone16$ozone_ppb)
X_cov = Kyy - t(Kxy) %*% solve(Kxx, Kxy)

# visualize
par(mfrow = c(1,3))
quilt.plot(cbind(ozone16$coords_lon, ozone16$coords_lat),
           ozone16$ozone_ppb, main = "ozone measurements"); US(add= T)

quilt.plot(cbind(health_sim$coords_y_lon, health_sim$coords_y_lat),
           X_mean, main = "health subjects, mean of exposure"); US(add= T)
points(cbind(ozone16$coords_lon, ozone16$coords_lat), pch = 17)

quilt.plot(cbind(health_sim$coords_y_lon, health_sim$coords_y_lat),
           sqrt(diag(X_cov)), main = "health subjects, sd of exposure"); US(add= T)
points(cbind(ozone16$coords_lon, ozone16$coords_lat), pch = 17)

# vecchia approximation
run_vecchia = vecchia_cov(X_cov, coords = cbind(health_sim$coords_y_lon, health_sim$coords_y_lat),
                          n.neighbors = 10)
Q_sparse = run_vecchia$Q
run_vecchia$cputime

# fit the model
fit_me = blogireg_me(Y = health_sim$Ybinary,
                X_mean = X_mean,
                X_prec = Q_sparse, # sparse precision matrix
                Z = health_sim$Z,
                nburn = 100,
                nsave = 1000,
                nthin = 1)
fit_me$cputime
summary(fit_me$posterior)
library(bayesplot)
bayesplot::mcmc_trace(fit_me$posterior)
}