Run Vecchia approximation given a covariance matrix

Given a multivariate normal (MVN) distribution with covariance matrix \(\Sigma\), this function finds a sparse precision matrix (inverse covariance) \(Q\) based on the Vecchia approximation (Vecchia 1988, Katzfuss and Guinness 2021), where \(N(\mu, Q^{-1})\) is the sparse MVN that approximates the original MVN \(N(\mu, \Sigma)\). The algorithm is based on the pseudocode 2 of Finley et al. (2019).

Usage

vecchia_cov(Sigma, coords, n.neighbors, ord = NULL, KLdiv = FALSE)

Arguments

Sigma

matrix<num>, n by n covariance matrix

coords

matrix<num>, n by 2 coordinate matrix for nearest neighborhood search

n.neighbors

integer, the number of nearest neighbors (k) to determine conditioning set of Vecchia approximation

ord

vector<int>, length n vector, ordering of data. If NULL, ordering based on the first coordinate will be used.

KLdiv

logical, If TRUE, return KL divergence \(D_{KL}(p || \tilde{p})\) where \(p\) is multivariate normal with original covariance matrix and \(\tilde{p}\) is the approximated multivariate normal with sparse precision matrix.

Value

list of the following:

Q

n by n sparse precision matrix in Matrix format

ord

ordering used for Vecchia approximation

cputime

time taken to run Vecchia approximation

KLdiv

(if KLdiv = TRUE) KL divergence \(D_{KL}(p || \tilde{p})\) where \(p\) is the multivariate normal with original covariance matrix and \(\tilde{p}\) is the approximated multivariate normal with a sparse precision matrix.

References

Vecchia, A. V. (1988). Estimation and model identification for continuous spatial processes. Journal of the Royal Statistical Society Series B: Statistical Methodology, 50(2), 297-312.

Katzfuss, M., & Guinness, J. (2021). A General Framework for Vecchia Approximations of Gaussian Processes. Statistical Science, 36(1).

Finley, A. O., Datta, A., Cook, B. D., Morton, D. C., Andersen, H. E., & Banerjee, S. (2019). Efficient algorithms for Bayesian nearest neighbor Gaussian processes. Journal of Computational and Graphical Statistics, 28(2), 401-414.

Zhang, L., (2020), public Github repository https://github.com/LuZhangstat/NNGP_STAN.

Examples

n = 1000
coords = cbind(runif(n), runif(n))
Sigma = fields::Exp.cov(coords, aRange = 1)
fit5 = vecchia_cov(Sigma, coords, n.neighbors = 5, KLdiv = TRUE)
fit5$KLdiv
#> [1] 1091.626
fit10 = vecchia_cov(Sigma, coords, n.neighbors = 10, KLdiv = TRUE)
fit10$KLdiv
#> [1] 917.0829